refactor(ir): move ConstrainedBoundary to ConstraintBuilder
