+++ ir/src/validation/mod.rs
@@ -1,7 +1,6 @@
 use super::{
-    constraints::{ConstrainedBoundary, ConstraintDomain},
-    Constant, IndexedTraceAccess, MatrixAccess, NamedTraceAccess, Symbol, SymbolType, TraceSegment,
-    VectorAccess, MIN_CYCLE_LENGTH,
+    constraints::ConstraintDomain, Constant, ConstrainedBoundary, IndexedTraceAccess, MatrixAccess,
+    NamedTraceAccess, Symbol, SymbolType, TraceSegment, VectorAccess, MIN_CYCLE_LENGTH,
 };
 
 mod error;
